requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.32.5
openai==0.28.1
dateparser==1.1.8
python-dateutil==2.8.2
schedule==1.2.0
beautifulsoup4==4.12.2
feedparser==6.0.12
dnspython==2.4.2
lxml==4.9.3
